the CO2 ppm was sended to the server

+++ 0.sourceCodes/0.finalPrograms/MG811-02282022/mg811library/mg811library.ino
@@ -18,7 +18,7 @@ void setup() {
 }
 
 void loop() {
-  int val = co2Sensor.read();
+  float val = co2Sensor.read();
   Serial.print("CO2 value: ");
   Serial.println(val);
